Remove several redundant includes and unused functions
Also fixed many signed/unsigned int inconsistencies.

NOTE: Certain parts of the code are littered with implicit conversions
between different sizes of signed and unsigned ints as well as doubles.
At some point we really need to refactor these and ensure consistency.
It is quite likely that this will blow up once systems get large
enough to start overflowing int or u_int.

+++ src/sym_table.c
@@ -394,7 +394,7 @@ static int resize_symtab(struct sym_table_head *hashtab, int size) {
       struct sym_table *entry = entries[i];
       entries[i] = entries[i]->next;
 
-      unsigned int hashval = hash(entry->name) & (size - 1);
+      unsigned long hashval = hash(entry->name) & (size - 1);
       entry->next = hashtab->entries[hashval];
       hashtab->entries[hashval] = entry;
     }
@@ -423,10 +423,10 @@ static void maybe_grow_symtab(struct sym_table_head *hashtab) {
 struct sym_table *store_sym(char const *sym, enum symbol_type_t sym_type,
                             struct sym_table_head *hashtab, void *data) {
   struct sym_table *sp;
-  unsigned hashval;
+  unsigned long hashval;
   void *vp = NULL;
   double *fp;
-  unsigned rawhash;
+  unsigned long rawhash;
 
   /* try to find sym in table */
   if ((sp = retrieve_sym(sym, hashtab)) == NULL) {
